optimizer generates decision tree

+++ util/src/main/java/jcog/net/MeshOptimize.java
@@ -28,8 +28,8 @@ public class MeshOptimize<X> extends Optimize<X> {
     }
 
     @Override
-    protected void onExperiment(double[] point, float score) {
+    protected void onExperiment(double[] point, double score) {
         super.onExperiment(point, score);
-        m.put(serial.incrementAndGet(), Floats.asList(ArrayUtils.add(Util.doubleToFloatArray(point), score)));
+        m.put(serial.incrementAndGet(), Floats.asList(ArrayUtils.add(Util.doubleToFloatArray(point), (float)score)));
     }
 }
